Switch to `yarn` to fix security vulnerabilities
- Replace all `npm` commands with `yarn`
The reason to do this switch is due to two reasons.
The first is that `npm install` completes in 1m30s, while
`npm install` completes in 50s, on average. The second is
explained in the point below. This runtime difference might appear
insignificant, but may improve development time over the long run.

- Remove `preinstall` script and change `resolutions` packages
The `preinstall` script is misleading and does not run before
dependencies are installed, and this is acknowledged in
https://github.com/npm/cli/issues/2660 . With the switch to
`yarn`, the `preinstall` script becomes obsolete as `yarn` will
take care of the `resolutions` without needing a script. In
addition, the cause of the security issues was misattributed to
the wrong dependency in the previous commit. Remove this and add
the true culprits.