Refactored code and added regex.
Removed telephone input for Signup.tsx and instead added first and last name input fields on the same line. Refactored the imports in index.tsx. Removed unwanted imports. Changed the background image: instead of using a possible copyrighted image, uploaded a new image created in photoshop labelled as `background.jpg`.

Made the AuthContainer more responsive as it changes based on the screen size. Implemented regex for the input fields that ensures that the information is not submitted until all the errors are resolved in the input fields. Such as confirming if the passwords matched each other and if the email belongs to the utoronto domain.