Setup email controllers: `emailService`
- Functions created for generating specific confirmation tokens
- Includes creating both types of tokens: password reset and confirmation tokens
- Setup SendGrid API and email format
- Add basic tests for the email controllers

The "token" that is generated will be a random alphanumeric string. When we store this in the db,
we append either `c:` or `r:` to the start so that we can check whether it should be used to reset a password, or to confirm the user account.

For example, if the token generated for email confirmation is `abcd`, we store `c:abcd` in the database for that user.

The emails that are sent to the user will be in the format `https://WEBSITE.com/c=<token>&e=<email>` for mail confirmation, or `https://WEBSITE.com/r=<token>&e=<email>` for password resets.