Upgrade login page design, fix resizing issues (#28)
* Upgrade login page design, fix resizing issues

- Use theming with createTheme instead of hardcoding colour values

- Use CssBaseline to provide normalized CSS stylings

- Fix ts error caused by conditional expect in App.test.tsx

- Use a hook to dynamically update signin and signup box sizes
Updates based on the size of the current viewport, so boxes can
adjust in realtime on screen size change.

- Allow scrolling to resolve issue of overflowing elements

- Add "Remember me" checkbox
This should be implemented in the backend in a future PR.

- Add "Forgot password?" link for password reset
This should be connected to the backend in a future PR.

- Update style of signup redirect and signin redirect buttons
Instead of using a Button style, use a Link component style.

- Update size of TextFields to be smaller
This will help with sizing errors and make things look more
compact.

- Send password mismatch to only confirm password field
The error does not need to be sent to both password fields and
also does not look as nice.

- Remove unnecessary TextField placeholders
Only email placeholder is really necessary to give an example of
a valid email address. It can be assumed that most people know how
to input their name, username, and password.

- Update background image and its css properties
Image needs to stretch to the full viewport.

- Use media query with breakpoint to hide "A place to connect"
This text does not fit and causing overflow when the screen size
is too small. Only make the text visible to those who have a
sufficiently large viewport or screen.

- Set overflow to hidden for Grid item containing Tabs
This should prevent the ripple effect, when clicking a tab, from
going past the rounded edges.

- Fix typos and sentence structure

* Fix ansi-html vulnerability

- Use ansi-html-community which patches this vulnerability

* Remove height calculation and move themes down level from App

* Moved themes to App.tsx and updated bg image

* Rearranged imports