Add ability to track which posts a user likes. (#64)
Add ability to track which posts a user likes.

Create a new model which tracks when a user likes/unlikes a post. Additionally, add logic to like
and unlike posts. Finally, each `Post` response includes the number of likes and whether or not the
current user liked the given post.

This is part of #38